.Net Standard Support
More updates to support .Net Standard 2.0
