+++ .nuget/NuGet.Config
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <configuration>
-  <solution>
-    <add key="disableSourceControlIntegration" value="true" />
-  </solution>
+  <packageSources>
+    <add key="nuget.org" value="https://api.nuget.org/v3/index.json" protocolVersion="3" />
+  </packageSources>
 </configuration>
