--- DapperExtensions.Net45/DapperExtensions.Net45.csproj
@@ -1,136 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{1656EC3C-7876-4C67-9650-C61ACB806D35}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>DapperExtensions</RootNamespace>
-    <AssemblyName>DapperExtensions</AssemblyName>
-    <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
-    <FileAlignment>512</FileAlignment>
-    <SolutionDir Condition="$(SolutionDir) == '' Or $(SolutionDir) == '*Undefined*'">..\..\basis\src\</SolutionDir>
-    <RestorePackages>true</RestorePackages>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType>pdbonly</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-  <ItemGroup>
-    <Reference Include="Dapper, Version=1.50.2.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\Dapper.1.50.2\lib\net45\Dapper.dll</HintPath>
-      <Private>True</Private>
-    </Reference>
-    <Reference Include="System" />
-    <Reference Include="System.Core" />
-    <Reference Include="System.Numerics" />
-    <Reference Include="System.Xml.Linq" />
-    <Reference Include="System.Data.DataSetExtensions" />
-    <Reference Include="Microsoft.CSharp" />
-    <Reference Include="System.Data" />
-    <Reference Include="System.Xml" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="..\DapperExtensions\DapperExtensions.cs">
-      <Link>DapperExtensions.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\DapperExtensionsConfiguration.cs">
-      <Link>DapperExtensionsConfiguration.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\DapperImplementor.cs">
-      <Link>DapperImplementor.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\Database.cs">
-      <Link>Database.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\GetMultiplePredicate.cs">
-      <Link>GetMultiplePredicate.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\GetMultipleResult.cs">
-      <Link>GetMultipleResult.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\Mapper\AutoClassMapper.cs">
-      <Link>Mapper\AutoClassMapper.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\Mapper\ClassMapper.cs">
-      <Link>Mapper\ClassMapper.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\Mapper\PluralizedAutoClassMapper.cs">
-      <Link>Mapper\PluralizedAutoClassMapper.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\Mapper\PropertyMap.cs">
-      <Link>Mapper\PropertyMap.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\Predicates.cs">
-      <Link>Predicates.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\Properties\AssemblyInfo.cs">
-      <Link>Properties\AssemblyInfo.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\ReflectionHelper.cs">
-      <Link>ReflectionHelper.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\Sql\DB2Dialect.cs">
-      <Link>Sql\DB2Dialect.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\Sql\MySqlDialect.cs">
-      <Link>Sql\MySqlDialect.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\Sql\OracleDialect.cs">
-      <Link>Sql\OracleDialect.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\Sql\PostgreSqlDialect.cs">
-      <Link>Sql\PostgreSqlDialect.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\Sql\SqlCeDialect.cs">
-      <Link>Sql\SqlCeDialect.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\Sql\SqlDialectBase.cs">
-      <Link>Sql\SqlDialectBase.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\Sql\SqlGenerator.cs">
-      <Link>Sql\SqlGenerator.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\Sql\SqliteDialect.cs">
-      <Link>Sql\SqliteDialect.cs</Link>
-    </Compile>
-    <Compile Include="..\DapperExtensions\Sql\SqlServerDialect.cs">
-      <Link>Sql\SqlServerDialect.cs</Link>
-    </Compile>
-    <Compile Include="DapperAsyncExtensions.cs" />
-    <Compile Include="DapperAsyncImplementor.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="packages.config" />
-  </ItemGroup>
-  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
-  <Import Project="$(SolutionDir)\.nuget\NuGet.targets" Condition="Exists('$(SolutionDir)\.nuget\NuGet.targets')" />
-  <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
-    <PropertyGroup>
-      <ErrorText>This project references NuGet package(s) that are missing on this computer. Enable NuGet Package Restore to download them.  For more information, see http://go.microsoft.com/fwlink/?LinkID=322105. The missing file is {0}.</ErrorText>
-    </PropertyGroup>
-    <Error Condition="!Exists('$(SolutionDir)\.nuget\NuGet.targets')" Text="$([System.String]::Format('$(ErrorText)', '$(SolutionDir)\.nuget\NuGet.targets'))" />
-  </Target>
-  <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
-       Other similar extension points exist, see Microsoft.Common.targets.
-  <Target Name="BeforeBuild">
-  </Target>
-  <Target Name="AfterBuild">
-  </Target>
-  -->
-</Project>
