--- DapperExtensions.Net45/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Dapper" version="1.50.2" targetFramework="net45" />
-</packages>
