+++ DapperExtensions.Test/Helpers/TestHelpers.cs
@@ -1,18 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data;
-using System.Data.SQLite;
-using System.Data.SqlClient;
-using System.Data.SqlServerCe;
-using System.IO;
-using System.Linq;
-using System.Threading;
-using Dapper;
-using DapperExtensions.Sql;
-using DapperExtensions.Test.IntegrationTests;
-using MySql.Data.MySqlClient;
-
-namespace DapperExtensions.Test.Helpers
+﻿namespace DapperExtensions.Test.Helpers
 {
     public static class TestHelpers
     {
