--- DapperExtensions/DapperExtensions.nuspec
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
-  <metadata schemaVersion="2">
-    <id>DapperExtensions</id>
-    <version>1.6.3</version>
-    <authors>Thad Smith, Page Brooks</authors>
-    <owners>Thad Smith, Page Brooks</owners>
-    <licenseUrl>http://www.apache.org/licenses/LICENSE-2.0</licenseUrl>
-    <projectUrl>https://github.com/tmsmith/Dapper-Extensions</projectUrl>
-    <!--<iconUrl>http://ICON_URL_HERE_OR_DELETE_THIS_LINE</iconUrl>-->
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>A small library that complements Dapper by adding basic CRUD operations (Get, Insert, Update, Delete) for your POCOs. For more advanced querying scenarios, Dapper Extensions provides a predicate system.</description>
-    <releaseNotes>
-      1.6.2 / 1.6.3
-      * Exposed Ignore keys on update in usages of Implementor
-      1.6.1
-      * Added Oracle Dialect to 4.5
-      1.6.0
-      * Added DB2 Dialect
-      * Added flag to ignore All Key fields when doing an update
-      * Added missing async/await methods
-      * Added unmap method to Mapper
-    </releaseNotes>
-    <tags>orm sql micro-orm Data Dapper</tags>
-	<dependencies>
-      <dependency id="Dapper" version="1.12" />
-    </dependencies>
-    <frameworkAssemblies>
-      <frameworkAssembly assemblyName="System.Core"/>
-      <frameworkAssembly assemblyName="System"/>
-      <frameworkAssembly assemblyName="System.Data"/>
-      <frameworkAssembly assemblyName="Microsoft.CSharp" targetFramework=".NETFramework4.0-Client, .NETFramework4.0" />
-    </frameworkAssemblies>
-  </metadata>
-  <files>
-  	<file src="bin\Release\DapperExtensions.dll" target="lib\net40" />
-    <file src="bin\Release\DapperExtensions.pdb" target="lib\net40" />
-
-    <file src="..\DapperExtensions.Net45\bin\Release\DapperExtensions.dll" target="lib\net45" />
-    <file src="..\DapperExtensions.Net45\bin\Release\DapperExtensions.pdb" target="lib\net45" />
-  </files>  
-</package>
