+++ DapperExtensions/DapperExtensionsConfiguration.cs
@@ -91,14 +91,14 @@ namespace DapperExtensions
             {
                 Type[] types = a.GetTypes();
                 return (from type in types
-                        let interfaceType = type.GetInterface(typeof(IClassMapper<>).FullName)
+                        let interfaceType = type.GetInterfaces().Where(x => x.Name == typeof(IClassMapper<>).Name && x.Namespace == typeof(IClassMapper<>).Namespace).FirstOrDefault()
                         where
                             interfaceType != null &&
                             interfaceType.GetGenericArguments()[0] == entityType
                         select type).SingleOrDefault();
             };
 
-            Type result = getType(entityType.Assembly);
+            Type result = getType(entityType.GetTypeInfo().Assembly);
             if (result != null)
             {
                 return result;
@@ -113,7 +113,7 @@ namespace DapperExtensions
                 }
             }
 
-            return getType(entityType.Assembly);
+            return getType(entityType.GetTypeInfo().Assembly);
         }
     }
 }
