+++ DapperExtensions/Sql/PostgreSqlDialect.cs
@@ -13,18 +13,18 @@ namespace DapperExtensions.Sql
         }
 
         public override string GetPagingSql(string sql, int page, int resultsPerPage, IDictionary<string, object> parameters)
-		{
-			int startValue = page * resultsPerPage;
-			return GetSetSql(sql, startValue, resultsPerPage, parameters);
-		}
-		
-		public override string GetSetSql(string sql, int pageNumber, int maxResults, IDictionary<string, object> parameters)
-		{
-			string result = string.Format("{0} LIMIT @maxResults OFFSET @pageStartRowNbr", sql);
-			parameters.Add("@maxResults", maxResults);
-			parameters.Add("@pageStartRowNbr", pageNumber * maxResults);
-			return result;
-		}
+        {
+            int startValue = page * resultsPerPage;
+            return GetSetSql(sql, startValue, resultsPerPage, parameters);
+        }
+
+        public override string GetSetSql(string sql, int firstResult, int maxResults, IDictionary<string, object> parameters)
+        {
+            string result = string.Format("{0} LIMIT @firstResult OFFSET @pageStartRowNbr", sql);            
+            parameters.Add("@firstResult", firstResult);
+            parameters.Add("@maxResults", maxResults);
+            return result;
+        }
 
         public override string GetColumnName(string prefix, string columnName, string alias)
         {
