+++ DapperExtensions/Sql/SqlDialectBase.cs
@@ -121,11 +121,7 @@ namespace DapperExtensions.Sql
 
         public virtual string QuoteString(string value)
         {
-            if (IsQuoted(value) || value == "*")
-            {
-                return value;
-            }
-            return string.Format("{0}{1}{2}", OpenQuote, value.Trim(), CloseQuote);
+            return IsQuoted(value) ? value : string.Format("{0}{1}{2}", OpenQuote, value.Trim(), CloseQuote);
         }
 
         public virtual string UnQuoteString(string value)
