--- DapperExtensions/buildPackage.cmd
@@ -1,1 +1,0 @@
-..\.nuget\nuget.exe pack DapperExtensions.nuspec /o ..\nuget
